Move audio device and recording settings to menu bar and restructure UI
