+++ src/main/resources/fxml/general.fxml
@@ -1,32 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.ComboBox?>
 <?import javafx.scene.control.Label?>
 <?import javafx.scene.control.Slider?>
-<?import javafx.scene.control.TextField?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.shape.SVGPath?>
 
-<AnchorPane id="control-pane" prefHeight="352.0" prefWidth="402.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sh.ball.gui.controller.GeneralController">
+<AnchorPane id="control-pane" prefHeight="324.0" prefWidth="364.0" xmlns="http://javafx.com/javafx/17" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sh.ball.gui.controller.GeneralController">
     <children>
-        <Button fx:id="chooseFileButton" layoutX="14.0" layoutY="53.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="114.0" text="Choose File" />
-        <Label fx:id="fileLabel" layoutX="143.0" layoutY="57.0" maxWidth="270.0" prefHeight="18.0" prefWidth="246.0" text="cube.obj" />
-        <Button fx:id="recordButton" layoutX="13.0" layoutY="142.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="114.0" text="Record" />
-        <Label fx:id="recordLabel" layoutX="13.0" layoutY="184.0" maxWidth="451.0" prefHeight="18.0" prefWidth="376.0" />
-        <Label id="frequency" fx:id="frequencyLabel" layoutX="13.0" layoutY="206.0" prefHeight="58.0" prefWidth="376.0" text="L/R Frequency:&#10; &#10;" />
-        <ComboBox fx:id="deviceComboBox" layoutX="14.0" layoutY="11.0" prefHeight="26.0" prefWidth="376.0" />
-        <Button fx:id="chooseFolderButton" layoutX="14.0" layoutY="91.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="114.0" text="Choose Folder" />
-        <Label fx:id="jkLabel" layoutX="143.0" layoutY="84.0" maxWidth="270.0" prefHeight="43.0" prefWidth="246.0" text="Use j and k (or MIDI Program Change) to cycle between files, or i to start playback" visible="false" wrapText="true" />
-        <TextField fx:id="recordTextField" disable="true" layoutX="296.0" layoutY="155.0" prefHeight="26.0" prefWidth="70.0" text="2.0" />
-        <Label fx:id="recordLengthLabel" disable="true" layoutX="286.0" layoutY="129.0" text="Record length (s)" />
-        <CheckBox fx:id="recordCheckBox" layoutX="154.0" layoutY="146.0" mnemonicParsing="false" text="Timed record" />
-        <Slider fx:id="octaveSlider" blockIncrement="1.0" layoutX="81.0" layoutY="271.0" majorTickUnit="1.0" max="5.0" min="-5.0" minorTickCount="0" prefHeight="42.0" prefWidth="280.0" showTickLabels="true" showTickMarks="true" snapToTicks="true" />
-        <SVGPath fx:id="octaveMidi" content="M20.15 8.26H22V15.74H20.15M13 8.26H18.43C19 8.26 19.3 8.74 19.3 9.3V14.81C19.3 15.5 19 15.74 18.38 15.74H13V11H14.87V13.91H17.5V9.95H13M10.32 8.26H12.14V15.74H10.32M2 8.26H8.55C9.1 8.26 9.41 8.74 9.41 9.3V15.74H7.59V10.15H6.5V15.74H4.87V10.15H3.83V15.74H2Z" fill="WHITE" layoutX="365.0" layoutY="267.0" pickOnBounds="true" />
-        <Label layoutX="12.0" layoutY="270.0" text="Octave" />
-      <Slider fx:id="micVolumeSlider" blockIncrement="1.0" layoutX="81.0" layoutY="310.0" majorTickUnit="1.0" max="10.0" prefHeight="42.0" prefWidth="280.0" showTickLabels="true" showTickMarks="true" value="1.0" />
-      <SVGPath fx:id="micVolumeMidi" content="M20.15 8.26H22V15.74H20.15M13 8.26H18.43C19 8.26 19.3 8.74 19.3 9.3V14.81C19.3 15.5 19 15.74 18.38 15.74H13V11H14.87V13.91H17.5V9.95H13M10.32 8.26H12.14V15.74H10.32M2 8.26H8.55C9.1 8.26 9.41 8.74 9.41 9.3V15.74H7.59V10.15H6.5V15.74H4.87V10.15H3.83V15.74H2Z" fill="WHITE" layoutX="365.0" layoutY="306.0" pickOnBounds="true" />
-      <Label layoutX="12.0" layoutY="309.0" text="Mic Volume" />
+        <Button fx:id="chooseFileButton" layoutX="14.0" layoutY="26.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="114.0" text="Choose File" />
+        <Label fx:id="fileLabel" layoutX="143.0" layoutY="30.0" maxWidth="270.0" prefHeight="18.0" prefWidth="210.0" text="cube.obj" />
+        <Label fx:id="recordLabel" layoutX="13.0" layoutY="304.0" maxWidth="451.0" prefHeight="18.0" prefWidth="343.0" />
+        <Label id="frequency" fx:id="frequencyLabel" layoutX="14.0" layoutY="141.0" prefHeight="58.0" prefWidth="343.0" text="L/R Frequency:&#10; &#10;" />
+        <Button fx:id="chooseFolderButton" layoutX="14.0" layoutY="76.0" mnemonicParsing="false" prefHeight="26.0" prefWidth="114.0" text="Choose Folder" />
+        <Label fx:id="jkLabel" layoutX="143.0" layoutY="69.0" maxWidth="270.0" prefHeight="70.0" prefWidth="210.0" text="Use j and k (or MIDI Program Change) to cycle between files, or i to start playback" visible="false" wrapText="true" />
+        <Slider fx:id="octaveSlider" blockIncrement="1.0" layoutX="82.0" layoutY="215.0" majorTickUnit="1.0" max="5.0" min="-5.0" minorTickCount="0" prefHeight="42.0" prefWidth="246.0" showTickLabels="true" showTickMarks="true" snapToTicks="true" />
+        <SVGPath fx:id="octaveMidi" content="M20.15 8.26H22V15.74H20.15M13 8.26H18.43C19 8.26 19.3 8.74 19.3 9.3V14.81C19.3 15.5 19 15.74 18.38 15.74H13V11H14.87V13.91H17.5V9.95H13M10.32 8.26H12.14V15.74H10.32M2 8.26H8.55C9.1 8.26 9.41 8.74 9.41 9.3V15.74H7.59V10.15H6.5V15.74H4.87V10.15H3.83V15.74H2Z" fill="WHITE" layoutX="333.0" layoutY="211.0" pickOnBounds="true" />
+        <Label layoutX="13.0" layoutY="214.0" text="Octave" />
+      <Slider fx:id="micVolumeSlider" blockIncrement="1.0" layoutX="82.0" layoutY="254.0" majorTickUnit="1.0" max="10.0" prefHeight="42.0" prefWidth="246.0" showTickLabels="true" showTickMarks="true" value="1.0" />
+      <SVGPath fx:id="micVolumeMidi" content="M20.15 8.26H22V15.74H20.15M13 8.26H18.43C19 8.26 19.3 8.74 19.3 9.3V14.81C19.3 15.5 19 15.74 18.38 15.74H13V11H14.87V13.91H17.5V9.95H13M10.32 8.26H12.14V15.74H10.32M2 8.26H8.55C9.1 8.26 9.41 8.74 9.41 9.3V15.74H7.59V10.15H6.5V15.74H4.87V10.15H3.83V15.74H2Z" fill="WHITE" layoutX="333.0" layoutY="250.0" pickOnBounds="true" />
+      <Label layoutX="13.0" layoutY="253.0" text="Mic Volume" />
     </children>
 </AnchorPane>
